+++ pydidas_plugins/proc_plugins/remove_1d_polynomial_bg.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2023, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -20,17 +22,17 @@ diffraction data.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
-__status__ = "Development"
+__status__ = "Production"
 __all__ = ["Remove1dPolynomialBackground"]
 
 import numpy as np
 from numpy.polynomial import Polynomial
 
+from pydidas.core import Dataset, Parameter, ParameterCollection, get_generic_parameter
 from pydidas.core.constants import PROC_PLUGIN, PROC_PLUGIN_INTEGRATED
-from pydidas.core import ParameterCollection, Parameter, Dataset, get_generic_parameter
 from pydidas.core.utils import process_1d_with_multi_input_dims
 from pydidas.plugins import ProcPlugin
 
@@ -56,7 +58,7 @@ DEFAULT_PARAMS_FOR_REMOVE1dPOLYBG = ParameterCollection(
         name="Polynomial fit order",
         tooltip=(
             "The polynomial order for the fit. This value "
-            "should typically not exceed3 or 4."
+            "should typically not exceed 3 or 4."
         ),
     ),
     Parameter(
@@ -212,7 +214,7 @@ class Remove1dPolynomialBackground(ProcPlugin):
         _tmpindices = np.where(_res[_local_res_min] <= 0.002)[0]
         _local_res_min = _local_res_min[_tmpindices]
 
-        _p_final = Polynomial.fit(_local_res_min, data[_local_res_min], 3)
+        _p_final = Polynomial.fit(_local_res_min, data[_local_res_min], self._fit_order)
 
         if self._kernel is not None:
             data = _raw